Split out functionality behind ModelTransformer.
