+++ lib/wings/model_transformer.rb
@@ -1,10 +1,5 @@
 # frozen_string_literal: true
 
-require 'wings/transformer_value_mapper'
-require 'wings/models/concerns/collection_behavior'
-require 'wings/hydra/works/models/concerns/work_valkyrie_behavior'
-require 'wings/hydra/works/models/concerns/file_set_valkyrie_behavior'
-
 module Wings
   ##
   # Transforms ActiveFedora models or objects into Valkyrie::Resource models or
@@ -22,7 +17,6 @@ module Wings
   #
   #   resource.alternate_ids # => [#<Valkyrie::ID:0x... id: 'an_identifier'>]
   #
-  # rubocop:disable Metrics/ClassLength
   class ModelTransformer
     ##
     # @!attribute [rw] pcdm_object
@@ -46,26 +40,12 @@ module Wings
     end
 
     ##
-    # @param reflections [Hash<Symbol, Object>]
-    #
-    # @return [Array<Symbol>]
-    def self.relationship_keys_for(reflections:)
-      relationships = reflections
-                      .keys
-                      .reject { |k| k.to_s.include?('id') }
-                      .map { |k| k.to_s.singularize + '_ids' }
-      relationships.delete('member_ids') # Remove here.  Members will be extracted as ordered_members in attributes method.
-      relationships.delete('ordered_member_proxy_ids') # This does not have a Valkyrie equivalent.
-      relationships
-    end
-
-    ##
     # Builds a `Valkyrie::Resource` equivalent to the `pcdm_object`
     #
     # @return [::Valkyrie::Resource] a resource mirroiring `pcdm_object`
     def build
       klass = ResourceClassCache.instance.fetch(pcdm_object) do
-        self.class.to_valkyrie_resource_class(klass: pcdm_object.class)
+        OrmConverter.to_valkyrie_resource_class(klass: pcdm_object.class)
       end
 
       mint_id unless pcdm_object.id
@@ -117,98 +97,6 @@ module Wings
       end
     end
 
-    ##
-    # Selects an existing base class for the generated valkyrie class
-    #
-    # @return [Class]
-    def self.base_for(klass:)
-      if klass == Hydra::AccessControls::Embargo
-        Hyrax::Embargo
-      elsif klass == Hydra::AccessControls::Lease
-        Hyrax::Lease
-      else
-        Hyrax::Resource
-      end
-    end
-
-    ##
-    # @param klass [String] an `ActiveFedora` model
-    #
-    # @return [Class] a dyamically generated `Valkyrie::Resource` subclass
-    #   mirroring the provided `ActiveFedora` model
-    #
-    # rubocop:disable Metrics/AbcSize
-    # rubocop:disable Metrics/MethodLength because metaprogramming a class
-    #   results in long methods
-    def self.to_valkyrie_resource_class(klass:)
-      relationship_keys = klass.respond_to?(:reflections) ? relationship_keys_for(reflections: klass.reflections) : []
-      relationship_keys.delete('member_ids')
-      relationship_keys.delete('member_of_collection_ids')
-      reflection_id_keys = klass.respond_to?(:reflections) ? klass.reflections.keys.select { |k| k.to_s.end_with? '_id' } : []
-
-      Class.new(base_for(klass: klass)) do
-        include Wings::CollectionBehavior if klass.included_modules.include?(Hyrax::CollectionBehavior)
-        include Wings::Works::WorkValkyrieBehavior if klass.included_modules.include?(Hyrax::WorkBehavior)
-        include Wings::Works::FileSetValkyrieBehavior if klass.included_modules.include?(Hyrax::FileSetBehavior)
-
-        # Based on Valkyrie implementation, we call Class.to_s to define
-        # the internal resource.
-        @internal_resource = klass.to_s
-
-        class << self
-          attr_reader :internal_resource
-        end
-
-        def self.to_s
-          internal_resource
-        end
-
-        klass.properties.each_key do |property_name|
-          attribute property_name.to_sym, ::Valkyrie::Types::String
-        end
-
-        relationship_keys.each do |linked_property_name|
-          attribute linked_property_name.to_sym, ::Valkyrie::Types::Set.of(::Valkyrie::Types::ID)
-        end
-
-        reflection_id_keys.each do |property_name|
-          attribute property_name, ::Valkyrie::Types::ID
-        end
-
-        # Defined after properties in case we have an `internal_resource` property.
-        # This may not be ideal, but based on my understanding of the `internal_resource`
-        # usage in Valkyrie, I'd rather keep synchronized the instance_method and class_method value for
-        # `internal_resource`
-        def internal_resource
-          self.class.internal_resource
-        end
-      end
-    end
-    # rubocop:enable Metrics/MethodLength
-    # rubocop:enable Metrics/AbcSize
-
-    class AttributeTransformer
-      def self.run(obj, keys)
-        # TODO: There is an open question about whether we want to treat all these relationships the same.  See Issue #3904.
-        attrs = keys.select { |k| k.to_s.end_with? '_ids' }.each_with_object({}) do |attr_name, mem|
-          mem[attr_name.to_sym] =
-            TransformerValueMapper.for(obj.try(attr_name)).result ||
-            TransformerValueMapper.for(attribute_ids_for(name: attr_name.chomp('_ids'), obj: obj)).result ||
-            TransformerValueMapper.for(attribute_ids_for(name: attr_name.chomp('_ids').pluralize, obj: obj)).result || []
-        end
-        keys.each_with_object(attrs) do |attr_name, mem|
-          next unless obj.respond_to?(attr_name) && !mem.key?(attr_name.to_sym)
-          mem[attr_name.to_sym] = TransformerValueMapper.for(obj.public_send(attr_name)).result
-        end
-      end
-
-      def self.attribute_ids_for(name:, obj:)
-        attribute_value = obj.try(name)
-        return if attribute_value.nil?
-        Array(attribute_value).map(&:id)
-      end
-    end
-
     private
 
       def mint_id
@@ -220,7 +108,7 @@ module Wings
       def attributes
         all_keys =
           pcdm_object.attributes.keys +
-          self.class.relationship_keys_for(reflections: pcdm_object.reflections)
+          OrmConverter.relationship_keys_for(reflections: pcdm_object.reflections)
 
         result = AttributeTransformer.run(pcdm_object, all_keys).merge(reflection_ids).merge(additional_attributes)
 
@@ -267,5 +155,4 @@ module Wings
         attrs[:lease] = Hyrax::Lease.new(**lease_attrs)
       end
   end
-  # rubocop:enable Style/ClassVars Metrics/ClassLength
 end
